refactor: change path of EmptyState component
