+++ src/pages/favorites.tsx
@@ -14,7 +14,7 @@ import { refreshToken } from "@/utils/refreshToken";
 import { getEnvVariables } from "@/utils/getEnvVeriables";
 import { ISignInData } from "@/types/apiSuperjobTypes";
 import { setToLocalStorage } from "@/utils/setToLocalStorage";
-import { EmptyState } from "@/EmptyState/EmptyState";
+import { EmptyState } from "@/components/EmptyState/EmptyState";
 
 interface IVacancies {
   page: number;
